backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.2.1
django-cors-headers==4.4.0
dj-database-url==2.2.0

celery==5.4.0
redis==5.0.8
django-celery-beat==2.6.0

python-dateutil==2.9.0.post0